Refactor(woorim/search): 라우팅 접근 파일로 index.js로 접근하도록 변경
