+++ app/src/apis/watch-list/watch-list.ctrl.js
@@ -1,4 +1,4 @@
-const WatchList = require("../../models/services/watchlist/WatchList");
+const WatchList = require("../../models/services/WatchList/WatchList");
 
 const output = {
   findAllById: async (req, res) => {
